[O2B-721] Prepare higher level selection model
