+++ lib/public/components/Filters/common/TagFilterModel.js
@@ -32,7 +32,7 @@ export class TagFilterModel extends Observable {
      *
      * @constructor
      *
-     * @param {PickerOption[]} defaultSelection the list of tags selected by default
+     * @param {SelectionOption[]} defaultSelection the list of tags selected by default
      * @param {boolean} defaultCollapse the default collapse status of the tags picker
      * @param {string} defaultCombinationOperator the default combination operator of the filter
      */
