+++ lib/public/components/common/picker/PickerModel.js
@@ -11,54 +11,24 @@
  * or submit itself to any jurisdiction.
  */
 
-import { Observable } from '/js/src/index.js';
-
-/**
- * @typedef PickerOption A picker option, with the actual value and its string representation
- * @property {number|string} value The id of the object this is used to see if it is checked.
- * @property {Component} [label] The representation of the option (if null, value is used as label)
- */
+import { SelectionModel } from '../selection/SelectionModel.js';
 
 /**
  * Model to handle the state of a tag picker
  */
-export class PickerModel extends Observable {
+export class PickerModel extends SelectionModel {
     /**
-     * Model's constructor, collapsed and without selection as a default (default are used as initial state and when the picker is reset)
+     * Constructor
      *
-     * @param {PickerOption[]} defaultSelection the default selection of tags
-     * @param {boolean} defaultCollapsed true if the tags must be collapsed as a default
+     * @param {SelectionOption[]} [defaultSelection=[]] the default selection
+     * @param {boolean} [defaultCollapsed=true] If true, the options list will be collapsed by default
      * @constructor
      */
     constructor(defaultSelection = [], defaultCollapsed = true) {
-        super();
-
-        /**
-         * @type {PickerOption[]}
-         * @private
-         */
-        this._defaultSelection = defaultSelection;
-
-        /**
-         * @type {PickerOption[]}
-         * @private
-         */
-        this._selectedOptions = [...defaultSelection];
+        super(defaultSelection);
 
         this._defaultCollapsed = defaultCollapsed;
         this._collapsed = defaultCollapsed;
-    }
-
-    /**
-     * States if the picker selection is exactly the default one
-     *
-     * @return {boolean} true if the selection is the default one
-     */
-    hasOnlyDefaultSelection() {
-        const { selected } = this;
-        const defaultSelection = [...new Set(this._defaultSelection.map(({ value }) => value))];
-
-        return selected.length === defaultSelection.length && selected.every((item) => defaultSelection.includes(item));
     }
 
     /**
@@ -77,41 +47,8 @@ export class PickerModel extends Observable {
      * @return {void}
      */
     reset() {
-        this._selectedOptions = [...this._defaultSelection];
+        super.reset();
         this._collapsed = this._defaultCollapsed;
-    }
-
-    /**
-     * States if the current picker is currently checked or not
-     *
-     * @param {PickerOption} option the option to check for selected state
-     * @return {boolean} true if the given picker is checked
-     */
-    isSelected(option) {
-        return this._selectedOptions.find((selectedOption) => option.value === selectedOption.value) !== undefined;
-    }
-
-    /**
-     * Remove the given option from the list of selected ones
-     *
-     * @param {PickerOption} option the option to deselect
-     * @return {void}
-     */
-    deselect(option) {
-        this.selectedOptions = this._selectedOptions.filter((selectedOption) => selectedOption.value !== option.value);
-    }
-
-    /**
-     * Add the given options to the list of selected ones
-     *
-     * @param {PickerOption} option the option to select
-     * @return {void}
-     */
-    select(option) {
-        if (!this.isSelected(option)) {
-            this._selectedOptions.push(option);
-        }
-        this.notify();
     }
 
     /**
@@ -121,35 +58,5 @@ export class PickerModel extends Observable {
      */
     get collapsed() {
         return this._collapsed;
-    }
-
-    /**
-     * Return the **values** of the currently selected options
-     *
-     * Do not use this getter to modify the selected list but use the `selected` setter to define the new selected list and to notify observers
-     *
-     * @return {string[]|number[]} the values of the selected options
-     */
-    get selected() {
-        return [...new Set(this._selectedOptions.map(({ value }) => value))];
-    }
-
-    /**
-     * Return the list of currently selected options
-     *
-     * @return {PickerOption[]} the currently selected options
-     */
-    get selectedOptions() {
-        return this._selectedOptions;
-    }
-
-    /**
-     * Define (overrides) the list of currently selected tags
-     *
-     * @param {PickerOption[]} selected the list of selected tags
-     */
-    set selectedOptions(selected) {
-        this._selectedOptions = selected;
-        this.notify();
     }
 }
