+++ lib/public/components/common/picker/picker.js
@@ -27,12 +27,12 @@ import { h, iconMinus, iconPlus } from '/js/src/index.js';
 /**
  * Returns the list of input components that correspond to the given options list
  *
- * @param {PickerOption[]} pickerOptions the list of options to display
- * @param {string} selector selector to customize input components selectors
  * @param {PickerModel} pickerModel the model containing the current inputs states
+ * @param {SelectionOption[]} pickerOptions the list of options to display
+ * @param {string} selector selector to customize input components selectors
  * @return {Component[]} the list of inputs
  */
-const mapOptionsToInput = (pickerOptions, selector, pickerModel) => pickerOptions.map((
+const mapOptionsToInput = (pickerModel, pickerOptions, selector) => pickerOptions.map((
     pickerOption,
     index,
 ) => h(`.form-check.${selector}-option`, [
@@ -59,7 +59,7 @@ const mapOptionsToInput = (pickerOptions, selector, pickerModel) => pickerOption
 /**
  * Generates a picker component
  *
- * @param {PickerOption[]} pickerOptions The available options of the picker
+ * @param {SelectionOption[]} pickerOptions The available options of the picker
  * @param {PickerModel} pickerModel The picker model.
  * @param {PickerConfiguration} [configuration] the picker's configuration
  * @returns {Component} A filterable collapsable picker.
@@ -67,7 +67,7 @@ const mapOptionsToInput = (pickerOptions, selector, pickerModel) => pickerOption
 export const picker = (pickerOptions, pickerModel, configuration) => {
     const { limit = 5, selector = 'picker', attributes = null, outlineSelection = false, placeHolder = null } = configuration || {};
 
-    const checkboxes = mapOptionsToInput(pickerOptions, selector, pickerModel);
+    const checkboxes = mapOptionsToInput(pickerModel, pickerOptions, selector);
 
     const selectedPills = pickerModel.selectedOptions.length > 0 && outlineSelection
         ? h(
