+++ lib/public/components/tag/tagToOption.js
@@ -15,6 +15,6 @@
  * Convert a tag to a picker option
  *
  * @param {Tag} tag the tag to convert
- * @return {PickerOption} the tag's option
+ * @return {SelectionOption} the tag's option
  */
 export const tagToOption = ({ text }) => ({ value: text });
